Resolved conflicts w master (main classes)
